requirements.txt
Flask
Flask-SocketIO
python-socketio[asyncio_server]
eventlet